% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_based.R
\name{CalculateSCCellProportions}
\alias{CalculateSCCellProportions}
\title{Calculate cell proportions based on single-cell data}
\usage{
CalculateSCCellProportions(sc.eset, subject.names, cell.types)
}
\arguments{
\item{sc.eset}{Expression Set with single-cell data}

\item{subject.names}{A character string. Name of phenoData attribute in
sc.eset that indicates individual ID.}

\item{cell.types}{A character string. Name of phenoData attribute in sc.eset
that indicates cell type}
}
\value{
sc.props Matrix. Cell proportions with number of cell types rows
  by number of individuals columns
}
\description{
Returns proportion of each cell type out of total cells for each individual
in the single-cell Expression Set
}
