% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_free.R
\name{CorTri}
\alias{CorTri}
\title{Correlate columns of data frame}
\usage{
CorTri(x, method = "pearson")
}
\arguments{
\item{x}{Data frame or matrix. Column vectors are correlated}

\item{method}{Character string. Name of method passed to cor. Pearson by default}
}
\value{
cors Numeric vector. Correlation coefficients of pairs
}
\description{
This function runs correlation between markers of a data frame or matrix, 
returning the values of the lower/upper triangular of the correlation matrix
in a vector.
}
