% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CountsToCPM}
\alias{CountsToCPM}
\title{Convert counts data in Expression Set to counts per million (CPM)}
\usage{
CountsToCPM(eset)
}
\arguments{
\item{eset}{Expression Set containing counts assay data.}
}
\value{
eset Expression Set containing CPM assay data
}
\description{
Convert counts data in Expression Set to counts per million (CPM)
}
