% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix.R
\name{dmix}
\alias{dmix}
\title{Evaluate a mixture density}
\usage{
dmix(x, f, params, wts, log = FALSE, ...)
}
\arguments{
\item{x}{Points at which the mixture should be evaluated.  If the density 
is multivariate, then each row of \code{x} should contain one set of 
points at which the mixture should be evaluated.}

\item{f}{Unnormalized density used in the mixture. The function \eqn{f} 
should include a arguments \code{x}, \code{params}, and \code{log}, the
last of which returns \eqn{log(f(x))}.}

\item{params}{Matrix in which each row contains parameters for \code{f}.  The 
number of rows in \code{params} should match the number of mixture
components to evaluate.}

\item{wts}{vector of weights for each mixture component}

\item{log}{TRUE to return the log of the mixture density}

\item{...}{additional parameters to be passed to \code{f}}
}
\description{
Evaluate a mixture density
}
