% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_based.R
\name{EstimatePCACellTypeProportions}
\alias{EstimatePCACellTypeProportions}
\title{Estimate cell type proportions using first PC of expression matrix}
\usage{
EstimatePCACellTypeProportions(x, weighted = FALSE, w = NULL)
}
\arguments{
\item{x}{A sample by gene bulk expression matrix. Genes should be marker
genes}

\item{weighted}{Boolean. If weighted=TRUE, multiply scaled gene expression by
gene weights}

\item{w}{Numeric vector. Weights of genes}
}
\value{
ret List. Attribute \strong{pcs} contains matrix of PCs, where PC1
  should be used as estimates for cell type abundances
  Attribute \strong{sdev} contains eigenvalues of eigendecomposition of
  var-covar matrix. The 1st eigenvalue should explain most of the variance.
  Attribute \strong{genes} contains names of genes.
}
\description{
Estimate cell type proportions using first PC of expression matrix
}
