% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FilterUnexpressedGenes}
\alias{FilterUnexpressedGenes}
\title{Remove genes in Expression Set with zero expression in all samples}
\usage{
FilterUnexpressedGenes(eset, verbose = TRUE)
}
\arguments{
\item{eset}{Expression Set}

\item{verbose}{Boolean. Print logging info}
}
\value{
eset Expression Set with zero expression genes removed
}
\description{
Remove genes in Expression Set with zero expression in all samples
}
