% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_free.R
\name{GetNumGenes}
\alias{GetNumGenes}
\title{Get number of genes to use with no weighted information}
\usage{
GetNumGenes(x, min.gene = 25, max.gene = 200)
}
\arguments{
\item{x}{Numeric Matrix. A sample by gene expression matrix containing the
marker genes.}

\item{min.gene}{Numeric. Minimum number of genes to consider as markers.}

\item{max.gene}{Numeric. Maximum number of genes to consider as markers.}
}
\value{
best.n Numeric. Number of genes to use
}
\description{
Get number of genes to use with no weighted information
}
