% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_based.R
\name{GetNumGenesWeighted}
\alias{GetNumGenesWeighted}
\title{Get number of genes to use with weighted PCA}
\usage{
GetNumGenesWeighted(x, w, min.gene = 25, max.gene = 200)
}
\arguments{
\item{x}{Numeric Matrix. A sample by gene expression matrix containing the
marker genes.}

\item{w}{Numeric Vector. The weights of the genes that correspond to the
columns of x.}

\item{min.gene}{Numeric. Minimum number of genes to consider as markers.}

\item{max.gene}{Numeric. Maximum number of genes to consider as markers.}
}
\value{
best.n Numeric. Number of genes to use
}
\description{
Get number of genes to use with weighted PCA
}
