% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_free.R
\name{GetNumPC}
\alias{GetNumPC}
\title{Return number of PCs to use for PCA CTP estimation}
\usage{
GetNumPC(cors, pct_thresh = 0.5, verbose = TRUE)
}
\arguments{
\item{cors}{List. Contains correlation coefficients.}

\item{pct_thresh}{Numeric. Threshold of percent of coefficients that are greater than 0.}

\item{verbose}{Boolean. Print log info.}
}
\value{
best_pc_n Numeric. The number of PCs to regress out to remove global covariates.
}
\description{
Return the minimum number of PCs to regress out based on 
CTP-CTP correlations. Given a list of vectors, return 
the number of PCs with less than pct_thresh percent of 
coefficients that are greater than 0.
}
