% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_based.R
\name{GetOverlappingGenes}
\alias{GetOverlappingGenes}
\title{Find overlapping genes in single-cell data, bulk data, and marker genes}
\usage{
GetOverlappingGenes(sc.eset, bulk.eset, markers, verbose)
}
\arguments{
\item{sc.eset}{Expression Set with single-cell data}

\item{bulk.eset}{Expression Set with bulk data}

\item{markers}{Character vector. List of relevant marker genes}

\item{verbose}{Boolean. Print logging info}
}
\value{
overlapping.genes Character vector. List of genes found in markers
  and both datasets.
}
\description{
Find overlapping genes in single-cell data, bulk data, and marker genes
}
