% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_based.R
\name{GetOverlappingSamples}
\alias{GetOverlappingSamples}
\title{Find overlapping samples in single-cell and bulk data}
\usage{
GetOverlappingSamples(sc.eset, bulk.eset, subject.names, verbose)
}
\arguments{
\item{sc.eset}{Expression Set with single-cell data}

\item{bulk.eset}{Expression Set with bulk data}

\item{subject.names}{A character string. Name of phenoData attribute in
sc.eset that indicates individual ID (that would be found in bulk.eset
if overlapping)}

\item{verbose}{Boolean. Print logging info}
}
\value{
samples A list with attributes \emph{overlapping} and
  \emph{remaining}. Each attribute refers to a character vector that lists
  the samples found in both datasets and samples found only in bulk,
  respectively
}
\description{
Find overlapping samples in single-cell and bulk data
}
