% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_based.R
\name{GetUniqueMarkers}
\alias{GetUniqueMarkers}
\title{Get unique markers present in only 1 cell type}
\usage{
GetUniqueMarkers(x, gene_col = "gene")
}
\arguments{
\item{x}{Data frame. Contains column with marker gene names}

\item{gene_col}{Character string. Name of the column that contains
the marker genes}
}
\value{
x Data frame. Markers with non-unique markers removed
}
\description{
Given a data frame of marker genes for cell types, 
returns a new data frame with non-unique markers removed.
}
