% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itx.R
\name{itx}
\alias{itx}
\title{Named inverse transformation functions}
\usage{
itx(x, link)
}
\arguments{
\item{x}{Values at which to evaluate the inverse link function}

\item{link}{Character vector specifying link function for which the 
inverse link function should be evaluated.  Supports \code{'identity'},
\code{'log'}, and \code{'logit'}.}
}
\description{
Evaluates the inverse of the named link function at the locations
\code{x}.
}
\examples{
itx(0, 'logit')

}
