% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logjac.R
\name{logjac}
\alias{logjac}
\title{Wrapper to abstractly evaluate log-Jacobian functions for transforms}
\usage{
logjac(x, link)
}
\arguments{
\item{x}{values at which to evaluate \eqn{J(x)}}

\item{link}{Character vector specifying link function for which the 
inverse link function should be evaluated.  Supports \code{'identity'},
\code{'log'}, and \code{'logit'}.}
}
\description{
Wrapper to abstractly evaluate log-Jacobian functions for transforms
}
\examples{
logjac(1, 'logit')

}
\seealso{
\code{\link{jac.log}}, \code{\link{jac.logit}}
}
