% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePars.R
\name{mergePars}
\alias{mergePars}
\title{Merge pre-computed components of f(theta1 | theta2, X)}
\usage{
mergePars(x, y)
}
\arguments{
\item{x}{Output from one of the parallel calls in wtdMix()}

\item{y}{Another output from one of the parallel calls in wtdMix()}
}
\description{
For use in the parallel call in wtdMix()
}
