% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_free.R
\name{NormalizeByPC}
\alias{NormalizeByPC}
\title{Regress out global expression PCs from bulk.}
\usage{
NormalizeByPC(bulk.eset, max_pc, genes_subset = NULL)
}
\arguments{
\item{bulk.eset}{ExpressionSet. Contains bulk expression data to normalize}

\item{max_pc}{Numeric. Maximum number of PCs to regress out}

\item{genes_subset}{Character vector. Vector of genes to subset the ExpressionSet bulk object}
}
\value{
List with each element containing an ExpressionSet object of the normalized
bulk gene expression data, starting with 0 (non-normalized) up to and including 
max_pc
}
\description{
This function normalizes bulk expression data by first calculating
global PCs and then regressing them out sequentially. If genes_subset 
is given, only normalizes and returns expression data for those listed 
genes, speeding up computation.
}
