% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_free.R
\name{ReferenceFreeDeconvolution}
\alias{ReferenceFreeDeconvolution}
\title{Performs reference-free deconvolution of bulk expression using marker genes}
\usage{
ReferenceFreeDeconvolution(bulk.eset, markers, ct_col = "cluster",
  gene_col = "gene", min_gene = 5, max_gene = 200,
  weighted = FALSE, w_col = "avg_logFC", unique_markers = TRUE,
  verbose = TRUE)
}
\arguments{
\item{bulk.eset}{Expression Set. Normalized bulk expression data.}

\item{markers}{Data frame with columns specifying cluster and gene,
and optionally a column for weights, typically the fold-change of the gene.
Important that the genes for each cell type are row-sorted by signficance.}

\item{weighted}{Boolean. Whether to use weights for gene prioritization}

\item{unique_markers}{Boolean. If TRUE, subset markers to include only genes 
that are markers for only one cell type}

\item{verbose}{Boolean. Whether to print log info during deconvolution.
Errors will be printed regardless.}

\item{ct.col}{Character string. Column name specifying cluster/cell type
corresponding to each marker gene in \strong{markers}.}

\item{gene.col}{Character string. Column name specifying gene names in
\strong{markers}.}

\item{min.gene}{Numeric. Min number of genes to use for each cell type.}

\item{max.gene}{Numeric. Max number of genes to use for each cell type.}

\item{w.col}{Character string. Column name for weights, such as "avg_logFC",
in \strong{markers}}
}
\value{
A List. Slot \strong{bulk.props} contains estimated relative cell
  type abundances. Slot \strong{var.explained} contains variance explained by
  first 20 PCs for cell type marker genes. Slot \strong{genes.used} contains
  vector of genes used for deconvolution.
}
\description{
Estimates relative abundances of cell types from PCA-based deconvolution.
Uses a list of marker genes to subset the expression data, and returns the 
first PC of each sub-matrix as the cell type fraction estimates.
Optionally, weights for each marker gene can be used to prioritize genes
that are highly expressed in the given cell type.
}
\details{
Note that this method expects the input bulk data to be normalized, unlike
the reference-based method.
}
