% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sFit.R
\name{sFit}
\alias{sFit}
\title{Fit a spatially varying coefficient model}
\usage{
sFit(x, coords, init, nSamples, thin = 1, rw.initsd = 0.1,
  inits = list(), C = 1, alpha = 0.44, priors = list(sigmasq =
  list(a = 2, b = 1), rho = list(L = 0, U = 1), nu = list(L = 0, U = 1)))
}
\description{
Fit a spatially varying coefficient model
}
