% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateBarcode}
\alias{simulateBarcode}
\title{Simulate barcode for decomposition illustration}
\usage{
simulateBarcode(index, individual, barcode.length)
}
\arguments{
\item{index}{Integer. Index of cell ID from 0 to barcode.length to the
fourth power. Will generate a unique nucleotide barcode for each
index.}

\item{individual}{Character. ID of individual that the cell is from.}

\item{barcode.length}{Integer. Length of nucleotide barcode.}
}
\value{
Simulated barcode for cell from an individual
}
\description{
Generates a nucleotide barcode similar to those generated by
10x chromium sequencing platforms for illustration purposes.
Generates barcode and individual ID separated by '-' delimiter.
}
