% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sKrig.R
\name{sKrig}
\alias{sKrig}
\title{Fit a spatially varying coefficient model}
\usage{
sKrig(x, sFit, coords.krig, coords = sFit$coords, burn = 0,
  ncores = 1)
}
\description{
Fit a spatially varying coefficient model
}
