% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx.R
\name{tx}
\alias{tx}
\title{Named transformation functions}
\usage{
tx(x, link)
}
\arguments{
\item{x}{Values at which to evaluate the link function}

\item{link}{Character vector specifying link function to evaluate.  Supports 
\code{'identity'}, \code{'log'}, and \code{'logit'}.}
}
\description{
Evaluates the named link function at the locations \code{x}.
}
\examples{
smolBayes:::tx(0.5, 'logit')

}
