% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wBuild.R
\name{wBuild}
\alias{wBuild}
\title{Derive parameters for building integration grids}
\usage{
wBuild(f, init, dim.theta2 = length(init), approx = "condgauss",
  link = rep("identity", length(init)), link.params = rep(list(NA),
  length(init)), optim.control = list(maxit = 5000, method = "BFGS"),
  ...)
}
\arguments{
\item{f}{function used to derive the weight function \eqn{w}.
\code{f} must be able to be called via \code{f(par, log, ...)}}

\item{init}{initial guess for mode of \code{f}.}

\item{dim.theta2}{\code{wBuild} assumes \code{par} is partitioned such that 
\code{par=c(theta1,theta2)}.  \code{dim.theta2} specifies the size of the 
partition.  The default is to assume that \code{f} is defined without a 
\code{theta1} component.}

\item{approx}{Style of approximation (i.e., \eqn{w}) to be created from mode 
of \code{f}.
\describe{
  \item{\code{'gaussian'}}{Gaussian approximation for \code{theta2} at 
    the mode of \code{f}. Assumes \code{f} is proportional to the marginal 
    posterior density for \code{theta2}.}
  \item{\code{'condgauss'}}{Gaussian approximation for \code{theta2} at 
    the mode of \code{f}.  The approximation is conditioned on the value of
    the mode for \code{theta1}. Assumes \code{f} is proportional to the 
    joint posterior density for \code{theta1,theta2.}}
  \item{\code{'condgauss-laplace'}}{Gaussian approximation for 
    \code{theta2} at the mode of \code{f}.  The approximation is 
    conditioned on a separate laplace approximation of the marginal 
    posterior mode for \code{theta1}.  Assumes \code{f} is proportional to 
    the joint posterior density for \code{theta1,theta2.}}
  \item{\code{'margauss'}}{Gaussian approximation for 
    \code{theta2} at the mode of \code{f}.  Assumes \code{f} is 
    proportional to the joint posterior density for \code{theta1,theta2.}, 
    then uses the marginal mean and covariance from the posterior's 
    gaussian approximation.}
}}

\item{link}{character vector that specifies transformations used during
optimization and integration of \eqn{f(\theta_2 | X)}.  While
\eqn{\theta_2} may be defined on arbitrary support, \code{wtdMix} performs
optimization and integration of \eqn{\theta_2} on an unconstrained support.
The \code{link} vector describes the transformations that must be applied
to each element of \eqn{\theta_2}.  Jacobian functions for the
transformations will automatically be added to the optimization and
integration routines. Currently supported link functions are \code{'log'},
\code{'logit'}, and \code{'identity'}.}

\item{link.params}{Optional list of additional parameters for link
functions.  For example, the logit function can be extended to allow
mappings to any closed interval.   There should be one list entry for each
link function.  Specify NA if no additional arguments are passed.}

\item{...}{additional arguments needed for function evaluation.}
}
\description{
Note: \eqn{w} is defined on the transformed scale, but for convenience 
\code{f} is defined on the original scale.
}
