% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wMix.R
\name{wMix}
\alias{wMix}
\title{Construct a weighted mixture object}
\usage{
wMix(f1, f2, w, f1.precompute = function(x, ...) {     x },
  spec = "ff", level = 2, c.int = NULL, c.level = 2,
  c.init = NULL, c.link = rep("identity", length(c.init)),
  c.link.params = rep(list(NA), length(c.init)),
  c.optim.control = list(maxit = 5000, method = "BFGS"), ncores = 1,
  quadError = TRUE, ...)
}
\description{
Construct a weighted mixture object
}
