% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtdMix.R
\name{wtdMix}
\alias{wtdMix}
\title{Assemble a weighted mixture of posteriors}
\usage{
wtdMix(f1, f1.precompute = function(x, ...) {     x }, f2,
  w = "direct", w.init, w.link = NULL, level = 2, w.control = NULL,
  ...)
}
\arguments{
\item{f1}{evaluates f(theta1 | X, theta2).  f1 must be defined as
f1 = function(theta1, params, log, ...).
\describe{
  \item{theta1}{a matrix of parameters at which to evaluate
    f(theta1 | X, theta2). each row should be one set of values at which
    the density should be evaluated}
  \item{param}{a vector of parameters needed to evaluate
    f(theta1 | X, theta2).  In most cases par will equal theta2, but in
    some cases, f(theta1 | X, theta2) depends on functions of theta2,
    which can be precomputed (to save time) as the weighted mixture
    approximation is constructed.}
  \item{log}{TRUE to return log(f(theta1 | X, theta2))}
  \item{...}{additional arguments needed for function evaluation}
}}

\item{f1.precompute}{function that pre-computes parameters for evaluating
f(theta1 | X, theta2).  param must be defined as param = function(theta2).}

\item{f2}{evaluates f(theta2 | X).  f2 must be defined as
f2 = function(theta2, log, ...).  The interpretation of these arguments
is the same as for f1}

\item{w}{TRUE if the last element in the vector returned by the param
function should be added to the log-weights before they are normalized
before they are normalized.  should be a function of theta2}

\item{w.init}{initial guess for mode of f(theta2 | X)}

\item{level}{accuracy level (typically number of grid points for the
underlying 1D quadrature rule) [description from mvQuad::createNIGrid]}

\item{...}{additional arguments to pass to f1, f2, param}

\item{f2.link}{character vector that specifies transformations used during
optimization and integration of f(theta2 | X).  while theta2 may be
defined on arbitrary support, \code{wtdMix} performs optimization and
integration of theta2 on an unconstrained support.  the \code{link}
vector describes the transformations that must be applied to each
element of theta2.  Jacobian functions for the transformations will
automatically be added to the optimization and integration routines.
currently supported link functions are 'log', 'logit', and 'identity'.}

\item{control}{limited list of options to pass to optim.  should be a list
of lists:
\describe{
  \item{w}{(default) list(method='BFGS', maxit=5e4)
  }
}}
}
\description{
Assemble a weighted mixture of posteriors
}
