% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kCompute.R
\name{kCompute}
\alias{kCompute}
\title{Compute integration constants for unnormalized densities}
\usage{
kCompute(f, init, method = "BFGS", maxit = 10000, level = 2,
  log = FALSE, ...)
}
\arguments{
\item{f}{Unnormalized density for which to compute a scaled integration
constant.  the function \eqn{f} should include an argument \code{log}, 
which returns \eqn{log(f(x))}.}

\item{init}{Initial guess for the density's mode}

\item{method}{method to be used to search for the density's mode}

\item{maxit}{maximum number of iterations \code{optim} should use in 
searching for the density's mode}

\item{level}{accuracy level (typically number of grid points for the 
underlying 1D quadrature rule) [description from mvQuad::createNIGrid]}

\item{log}{TRUE to return log of integration constant}

\item{...}{additional arguments to pass to \code{f}}
}
\description{
Compute integration constants for unnormalized densities
}
