% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx.R
\name{tx}
\alias{tx}
\title{Named transformation functions}
\usage{
tx(x, link, linkparams)
}
\arguments{
\item{x}{Values at which to evaluate the link function}

\item{link}{Character vector specifying link function to evaluate.  Supports 
\code{'identity'}, \code{'log'}, and \code{'logit'}.}

\item{linkparams}{Optional list of additional parameters  for link functions.
For example, the logit function can be extended to allow mappings to any 
closed interval.  There should be one list entry for each link function.  
Specify NA if defaults should be used.}
}
\description{
Evaluates the named link function at the locations \code{x}.
}
\examples{
bisque:::tx(0.5, 'logit', list(NA))

}
